Initial Commit
Hello! If you're reading this then it Probably means this project is officially on GitHub. It's smol but it can be expanded upon. As of now there's currently 1 AI with a couple of side features. This entity is called "Warden" It's the Warden from Minecraft obviously. This project is open for anyone to come in and make changes. however your changes will be reviewed if you open a pull request. this is just here to make sure someone doesn't bomb the whole project qwq. anyways welcome >:3